Move maintenance/config/user operations to script endpoint
- Avoids the situation where actions can't be performed if the plugin can't start
  e.g. if access_manager.json is corrupt
