--- resources/lib/youtube_plugin/refresh.py
@@ -1,16 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-
-    Copyright (C) 2018-2018 plugin.video.youtube
-
-    SPDX-License-Identifier: GPL-2.0-only
-    See LICENSES/GPL-2.0-only for more information.
-"""
-
-from __future__ import absolute_import, division, unicode_literals
-
-import xbmc
-
-
-if __name__ == '__main__':
-    xbmc.executebuiltin("Container.Refresh")
