+++ resources/lib/youtube_plugin/youtube/provider.py
@@ -11,9 +11,7 @@
 from __future__ import absolute_import, division, unicode_literals
 
 import json
-import os
 import re
-import socket
 from base64 import b64decode
 
 from .client import YouTube
@@ -33,17 +31,19 @@ from .helper import (
 )
 from .youtube_exceptions import InvalidGrant, LoginException
 from ..kodion import AbstractProvider, RegisterProviderPath
-from ..kodion.compatibility import xbmcaddon, xbmcvfs
 from ..kodion.constants import (
     ADDON_ID,
-    DATA_PATH,
-    TEMP_PATH,
     content,
     paths,
 )
-from ..kodion.items import DirectoryItem, NewSearchItem, SearchItem, menu_items
-from ..kodion.network import get_client_ip_address, is_httpd_live
-from ..kodion.utils import find_video_id, rm_dir, strip_html_from_text
+from ..kodion.items import (
+    DirectoryItem,
+    NewSearchItem,
+    SearchItem,
+    UriItem,
+    menu_items,
+)
+from ..kodion.utils import find_video_id, strip_html_from_text
 
 
 class Provider(AbstractProvider):
@@ -677,109 +677,9 @@ class Provider(AbstractProvider):
     @RegisterProviderPath('^/users/(?P<action>[^/]+)/$')
     def _on_users(self, context, re_match):
         action = re_match.group('action')
-        refresh = context.get_param('refresh')
-
-        localize = context.localize
-        access_manager = context.get_access_manager()
-        ui = context.get_ui()
-
-        def select_user(reason, new_user=False):
-            current_users = access_manager.get_users()
-            current_user = access_manager.get_current_user()
-            usernames = []
-            for user, details in sorted(current_users.items()):
-                username = details.get('name') or localize('user.unnamed')
-                if user == current_user:
-                    username = '> ' + ui.bold(username)
-                if details.get('access_token') or details.get('refresh_token'):
-                    username = ui.color('limegreen', username)
-                usernames.append(username)
-            if new_user:
-                usernames.append(ui.italic(localize('user.new')))
-            return ui.on_select(reason, usernames), sorted(current_users.keys())
-
-        def add_user():
-            results = ui.on_keyboard_input(localize('user.enter_name'))
-            if results[0] is False:
-                return None, None
-            new_username = results[1].strip()
-            if not new_username:
-                new_username = localize('user.unnamed')
-            return access_manager.add_user(new_username)
-
-        def switch_to_user(user):
-            access_manager.set_user(user, switch_to=True)
-            ui.show_notification(
-                localize('user.changed') % access_manager.get_username(user),
-                localize('user.switch')
-            )
-            self.get_resource_manager(context).clear()
-            if refresh:
-                ui.refresh_container()
-
-        if action == 'switch':
-            result, user_index_map = select_user(localize('user.switch'),
-                                                 new_user=True)
-            if result == -1:
-                return True
-            if result == len(user_index_map):
-                user, _ = add_user()
-            else:
-                user = user_index_map[result]
-
-            if user is not None and user != access_manager.get_current_user():
-                switch_to_user(user)
-
-        elif action == 'add':
-            user, details = add_user()
-            if user is not None:
-                result = ui.on_yes_no_input(
-                    localize('user.switch'),
-                    localize('user.switch.now') % details.get('name')
-                )
-                if result:
-                    switch_to_user(user)
-
-        elif action == 'remove':
-            result, user_index_map = select_user(localize('user.remove'))
-            if result == -1:
-                return True
-
-            user = user_index_map[result]
-            username = access_manager.get_username(user)
-            if ui.on_remove_content(username):
-                access_manager.remove_user(user)
-                if user == 0:
-                    access_manager.add_user(username=localize('user.default'),
-                                            user=0)
-                if user == access_manager.get_current_user():
-                    access_manager.set_user(0, switch_to=True)
-                ui.show_notification(localize('removed') % username,
-                                     localize('remove'))
-
-        elif action == 'rename':
-            result, user_index_map = select_user(localize('user.rename'))
-            if result == -1:
-                return True
-
-            user = user_index_map[result]
-            old_username = access_manager.get_username(user)
-            results = ui.on_keyboard_input(localize('user.enter_name'),
-                                           default=old_username)
-            if results[0] is False:
-                return True
-            new_username = results[1].strip()
-            if not new_username:
-                new_username = localize('user.unnamed')
-            if old_username == new_username:
-                return True
-
-            if access_manager.set_username(user, new_username):
-                ui.show_notification(localize('renamed') % (old_username,
-                                                            new_username),
-                                     localize('rename'))
-
-        return True
+        return UriItem('{addon},users/{action}'.format(
+            addon=ADDON_ID, action=action
+        ))
 
     @RegisterProviderPath('^/sign/(?P<mode>[^/]+)/$')
     def _on_sign(self, context, re_match):
@@ -894,61 +794,12 @@ class Provider(AbstractProvider):
         result.extend(v3.response_to_items(self, context, json_data))
         return result
 
-    @RegisterProviderPath('^/config/(?P<switch>[^/]+)/$')
+    @RegisterProviderPath('^/config/(?P<action>[^/]+)/$')
     def configure_addon(self, context, re_match):
-        switch = re_match.group('switch')
-        localize = context.localize
-        settings = context.get_settings()
-        ui = context.get_ui()
-
-        if switch == 'youtube':
-            context.addon().openSettings()
-            ui.refresh_container()
-        elif switch == 'isa':
-            if context.use_inputstream_adaptive():
-                xbmcaddon.Addon(id='inputstream.adaptive').openSettings()
-            else:
-                settings.set_bool('kodion.video.quality.isa', False)
-        elif switch == 'subtitles':
-            yt_language = settings.get_string('youtube.language', 'en-US')
-            sub_setting = settings.subtitle_languages()
-
-            if yt_language.startswith('en'):
-                sub_opts = [localize('none'),
-                            localize('prompt'),
-                            localize('subtitles.with_fallback') % ('en', 'en-US/en-GB'),
-                            yt_language,
-                            '%s (%s)' % (yt_language, localize('subtitles.no_auto_generated'))]
-
-            else:
-                sub_opts = [localize('none'),
-                            localize('prompt'),
-                            localize('subtitles.with_fallback') % (yt_language, 'en'),
-                            yt_language,
-                            '%s (%s)' % (yt_language, localize('subtitles.no_auto_generated'))]
-
-            sub_opts[sub_setting] = ui.bold(sub_opts[sub_setting])
-
-            result = ui.on_select(localize('subtitles.language'), sub_opts)
-            if result > -1:
-                settings.set_subtitle_languages(result)
-
-            result = ui.on_yes_no_input(
-                localize('subtitles.download'),
-                localize('subtitles.download.pre')
-            )
-            if result > -1:
-                settings.set_subtitle_download(result == 1)
-        elif switch == 'listen_ip':
-            local_ranges = ('10.', '172.16.', '192.168.')
-            addresses = [iface[4][0]
-                         for iface in socket.getaddrinfo(socket.gethostname(), None)
-                         if iface[4][0].startswith(local_ranges)]
-            addresses += ['127.0.0.1', '0.0.0.0']
-            selected_address = ui.on_select(localize('select.listen.ip'), addresses)
-            if selected_address != -1:
-                settings.set_httpd_listen(addresses[selected_address])
-        return False
+        action = re_match.group('action')
+        return UriItem('{addon},config/{action}'.format(
+            addon=ADDON_ID, action=action
+        ))
 
     # noinspection PyUnusedLocal
     @RegisterProviderPath('^/my_subscriptions/filter/$')
@@ -992,101 +843,41 @@ class Provider(AbstractProvider):
                 ui.show_notification(message=message)
         ui.refresh_container()
 
-    @RegisterProviderPath('^/maintain/(?P<maint_type>[^/]+)/(?P<action>[^/]+)/$')
+    @RegisterProviderPath('^/maintenance/(?P<action>[^/]+)/(?P<target>[^/]+)/$')
     def maintenance_actions(self, context, re_match):
-        maint_type = re_match.group('maint_type')
+        target = re_match.group('target')
         action = re_match.group('action')
+
+        if action != 'reset':
+            return UriItem('{addon},maintenance/{action}/{target}'.format(
+                addon=ADDON_ID, action=action, target=target
+            ))
 
         ui = context.get_ui()
         localize = context.localize
 
-        if action == 'clear':
-            if maint_type == 'function_cache':
-                if ui.on_remove_content(localize('cache.function')):
-                    context.get_function_cache().clear()
-                    ui.show_notification(localize('succeeded'))
-            elif maint_type == 'data_cache':
-                if ui.on_remove_content(localize('cache.data')):
-                    context.get_data_cache().clear()
-                    ui.show_notification(localize('succeeded'))
-            elif maint_type == 'search_cache':
-                if ui.on_remove_content(localize('search.history')):
-                    context.get_search_history().clear()
-                    ui.show_notification(localize('succeeded'))
-            elif maint_type == 'playback_history' and ui.on_remove_content(localize('playback.history')):
-                context.get_playback_history().clear()
+        if (target == 'access_manager' and ui.on_yes_no_input(
+            context.get_name(), localize('reset.access_manager.confirm')
+        )):
+            try:
+                context.get_function_cache().clear()
+                access_manager = context.get_access_manager()
+                client = self.get_client(context)
+                if access_manager.has_refresh_token():
+                    refresh_tokens = access_manager.get_refresh_token()
+                    for refresh_token in set(refresh_tokens.split('|')):
+                        try:
+                            client.revoke(refresh_token)
+                        except:
+                            pass
+                self.reset_client()
+                access_manager.update_access_token(access_token='',
+                                                   refresh_token='')
+                ui.refresh_container()
                 ui.show_notification(localize('succeeded'))
-        elif action == 'reset':
-            if maint_type == 'access_manager' and ui.on_yes_no_input(context.get_name(), localize('reset.access_manager.confirm')):
-                try:
-                    context.get_function_cache().clear()
-                    access_manager = context.get_access_manager()
-                    client = self.get_client(context)
-                    if access_manager.has_refresh_token():
-                        refresh_tokens = access_manager.get_refresh_token().split('|')
-                        for refresh_token in set(refresh_tokens):
-                            try:
-                                client.revoke(refresh_token)
-                            except:
-                                pass
-                    self.reset_client()
-                    access_manager.update_access_token(access_token='', refresh_token='')
-                    ui.refresh_container()
-                    ui.show_notification(localize('succeeded'))
-                except:
-                    ui.show_notification(localize('failed'))
-        elif action == 'delete':
-            _maint_files = {'function_cache': 'cache.sqlite',
-                            'search_cache': 'search.sqlite',
-                            'data_cache': 'data_cache.sqlite',
-                            'playback_history': 'playback_history',
-                            'settings_xml': 'settings.xml',
-                            'api_keys': 'api_keys.json',
-                            'access_manager': 'access_manager.json',
-                            'temp_files': TEMP_PATH}
-            _file = _maint_files.get(maint_type)
-            succeeded = False
-
-            if not _file:
-                return
-
-            data_path = xbmcvfs.translatePath(DATA_PATH)
-            if 'sqlite' in _file:
-                _file_w_path = os.path.join(data_path, 'kodion', _file)
-            elif maint_type == 'temp_files':
-                _file_w_path = _file
-            elif maint_type == 'playback_history':
-                _file = ''.join((
-                    context.get_access_manager().get_current_user_id(),
-                    '.sqlite'
-                ))
-                _file_w_path = os.path.join(data_path, 'playback', _file)
-            else:
-                _file_w_path = os.path.join(data_path, _file)
-
-            if not ui.on_delete_content(_file):
-                return
-
-            if maint_type == 'temp_files':
-                succeeded = rm_dir(_file_w_path)
-
-            elif _file_w_path:
-                succeeded = xbmcvfs.delete(_file_w_path)
-
-            if succeeded:
-                ui.show_notification(localize('succeeded'))
-            else:
+            except:
                 ui.show_notification(localize('failed'))
 
-        elif action == 'install' and maint_type == 'inputstreamhelper':
-            if context.get_system_version().compatible(17):
-                try:
-                    xbmcaddon.Addon('script.module.inputstreamhelper')
-                    ui.show_notification(localize('inputstreamhelper.is_installed'))
-                except RuntimeError:
-                    context.execute('InstallAddon(script.module.inputstreamhelper)')
-            else:
-                ui.show_notification(localize('requires.krypton'))
 
     # noinspection PyUnusedLocal
     @RegisterProviderPath('^/api/update/$')
@@ -1141,20 +932,6 @@ class Provider(AbstractProvider):
                 log_list.append('Secret')
             ui.show_notification(localize('api.personal.failed') % ', '.join(missing_list))
             context.log_debug('Failed to enable personal API keys. Missing: %s' % ', '.join(log_list))
-
-    # noinspection PyUnusedLocal
-    @RegisterProviderPath('^/show_client_ip/$')
-    def show_client_ip(self, context, re_match):
-        port = context.get_settings().httpd_port()
-
-        if is_httpd_live(port=port):
-            client_ip = get_client_ip_address(port=port)
-            if client_ip:
-                context.get_ui().on_ok(context.get_name(), context.localize('client.ip') % client_ip)
-            else:
-                context.get_ui().show_notification(context.localize('client.ip.failed'))
-        else:
-            context.get_ui().show_notification(context.localize('httpd.not.running'))
 
     # noinspection PyUnusedLocal
     def on_playback_history(self, context, re_match):
